==> notebooks/03_Renaming_DNR_PlotClips.ipynb <==
In this notebook, we update the names of acquisitions from the DNR lidar portal to be consistent with the names that will come from the NOAA lidar server.

import pandas as pd
import os
import glob
import numpy as np

dnr_clips = pd.DataFrame(
    glob.glob('../data/raw/lidar/dnr_portal/dnrplot_clips/hectare_clips_epsg2927/*.laz') + \
    glob.glob('../data/raw/lidar/dnr_portal/dnrplot_clips/plot_clips_epsg2927/*.laz'),
    columns=['path'])
dnr_clips['basename'] = dnr_clips['path'].apply(lambda x: os.path.basename(x))
dnr_clips['dnr_acq_name'] = dnr_clips['basename'].apply(lambda x: '_'.join(x.split('_')[1:]).split('.laz')[0])
dnr_clips.head()

len(dnr_clips)

dnr_noaa_lookup = pd.read_csv('../data/raw/lidar/dnr_noaa_crosswalk.csv').set_index('dnr_name')
dnr_noaa_lookup.head()

noaa_name_lookup = pd.read_csv('../data/raw/lidar/noaa_tiles/noaa_acq_name_lookup.csv').set_index('noaa_id')
noaa_name_lookup.head()

dnr_clips.loc[dnr_clips.dnr_acq_name.isin(dnr_noaa_lookup.index.values)].head(10)

def rename(row):
    try:
        noaa_id = dnr_noaa_lookup.loc[row['dnr_acq_name']]
        new_name = noaa_name_lookup.loc[noaa_id].values[0][0]
        return new_name
    except KeyError:
        new_name = row['dnr_acq_name']
        new_name = new_name[:-5] + '_' + new_name[-4:]
        return new_name

dnr_clips['new_name'] = dnr_clips.apply(lambda x: rename(x), axis=1)

dnr_clips['new_path'] = dnr_clips.apply(lambda x: x['path'].replace(x['dnr_acq_name'], x['new_name']), axis=1)

print('Renaming {} files'.format(len(dnr_clips)))
for i, row in dnr_clips.iterrows():
    os.rename(row.path, row.new_path)
    print(i, end='.')
