==> notebooks/05_Reprojecting_NOAA_Tiles.ipynb <==
In this notebook, we reproject all point clouds from the geographic coordinate system they come with from the NOAA lidar server over to UTM. We will need to get the NOAA tiles into a projected coordinate system before we can clip plots from them.

import geopandas as gpd
import pandas as pd
import os
import glob
import subprocess
import dask
from dask.distributed import Client, LocalCluster, progress

We will need to know which UTM zone each tile is in.

If you haven't created a tile index already, you can do so like this:

TILE_DIR = '../data/raw/lidar/noaa_tiles/orig_laz/'
tiles = glob.glob(os.path.join(TILE_DIR, '*.laz'))

@dask.delayed
def get_boundary(infile, odir):
    proc = subprocess.run(['wine', '/storage/lidar/LAStools/bin/lasboundary.exe',
                           '-i', infile,
                           '-odir', odir,
                           '-oshp',
                           '-use_bb',
                           '-labels'], 
                          stderr=subprocess.PIPE,
                          stdout=subprocess.PIPE)
#     print(outfile)
    return proc

cluster=LocalCluster(scheduler_port=7001, diagnostics_port=7002)
c = Client(cluster)

res = c.persist([get_boundary(t, odir=TILE_DIR) for t in tiles])
progress(res)

# c.cancel(res)

# c.close()
# cluster.close()

# gather coverages of each tile
shps = glob.glob(os.path.join(TILE_DIR, '*.shp'))
len(shps)

# read in all the shapefiles of tile boudnaries into a list
gdfs = [gpd.read_file(shp) for shp in shps]
# concatentate them into a single geodataframe
gdf = pd.concat(gdfs, axis=0, ignore_index=True)
gdf.info()

gdf.crs

tile_idx = gdf.to_crs({'init':'epsg:4326'})
tile_idx.to_file('../data/raw/lidar/noaa_tiles/noaa_tileindex.shp')

Now we'll grab our UTM zones and do some spatial joins to assign each lidar tile to a UTM zone.

UTM_10 = '../data/external/utm_zone10_epsg4326.shp'
UTM_11 = '../data/external/utm_zone11_epsg4326.shp'
utm_10 = gpd.read_file(UTM_10)
utm_11 = gpd.read_file(UTM_11)
utm_10.info()

utm10_tiles = gpd.sjoin(tile_idx, utm_10)
utm11_tiles = gpd.sjoin(tile_idx, utm_11)
len(tile_idx), len(utm10_tiles), len(utm11_tiles)

utm_lookup = pd.concat((utm10_tiles, utm11_tiles),
                       axis=0, ignore_index=True)[['file_name', 'ZONE']].set_index('file_name')
utm_lookup['epsg'] = utm_lookup.ZONE.apply(lambda x: 6339 if x == '10' else 6340)
utm_lookup.head()

@dask.delayed
def reproject(infile, odir):
    basename = os.path.basename(infile)
    outfile = os.path.join(odir, basename)
    
    target_epsg = utm_lookup.loc[basename]['epsg']
    
    proc = subprocess.run(['wine', '/storage/lidar/LAStools/bin/las2las.exe',
                           '-i', infile,
                           '-o', outfile,
                           '-target_epsg', str(target_epsg)],
                          stderr=subprocess.PIPE,
                          stdout=subprocess.PIPE)
    return proc

ODIR = '../data/raw/lidar/noaa_tiles/utm_laz/'
res = c.persist([reproject(t, odir=ODIR) for t in tiles])
progress(res)

# c.close()
# cluster.close()
