==> notebooks/06_Clipping Plots from NOAA Tiles.ipynb <==
In this notebook, we will clip 1-hectare and smaller plot-sized excerpts from lidar point cloud tiles provided by NOAA. We will update the names of these acquisitions for consistency with other data sources we're combining them with (instead of the NOAA ID #s)

import geopandas as gpd
import pandas as pd
import os
import glob
import subprocess
import dask
from dask.distributed import LocalCluster, Client, progress

PLOTS = '../data/processed/blm_usfs_wadnr_plots_1ha.shp'
plots = gpd.read_file(PLOTS)
plots['geometry'] = plots.centroid
plots.info()

plots.crs

TILE_IDX = '../data/raw/lidar/noaa_tiles/noaa_tileindex.shp'
tile_idx = gpd.read_file(TILE_IDX)
tile_idx.info()

tile_idx.crs

tile_idx['noaa_id'] = tile_idx.file_name.apply(lambda x: x.split('_')[0].split('NOAA')[-1]).astype(int)
tile_idx = tile_idx.to_crs({'init':'epsg:4326'})

NOAA_NAMES = '../data/raw/lidar/noaa_tiles/noaa_acq_name_lookup.csv'
noaa_name_lookup = pd.read_csv(NOAA_NAMES).set_index('noaa_id')
noaa_name_lookup.head()

# assign acquisition names to noaa tiles based on tile_ids
tile_idx['acq_name'] = tile_idx.noaa_id.apply(lambda x: noaa_name_lookup.loc[x])

UTM_10 = '../data/external/utm_zone10_epsg4326.shp'
UTM_11 = '../data/external/utm_zone11_epsg4326.shp'
utm_10 = gpd.read_file(UTM_10)
utm_11 = gpd.read_file(UTM_11)
utm_zones = pd.concat((utm_10, utm_11),
                      axis=0,
                      ignore_index=True)[['geometry', 'ZONE']]
utm_zones.info()

plot_utm = gpd.sjoin(plots, utm_zones)[['uuid', 'ZONE', 'source', 'geometry']].set_index('uuid')
plot_utm['epsg'] = plot_utm.ZONE.apply(lambda x: 6339 if (x == '10') else 6340)
len(plot_utm)

# define the utm coordinates of the centroid of each plot in that zone
for epsg in [6339, 6340]:
    plot_utm.loc[plot_utm.epsg == epsg, 'utm_x'] = plot_utm.loc[plot_utm.epsg == epsg].to_crs({'init': 'epsg:{}'.format(epsg)}).centroid.x
    plot_utm.loc[plot_utm.epsg == epsg, 'utm_y'] = plot_utm.loc[plot_utm.epsg == epsg].to_crs({'init': 'epsg:{}'.format(epsg)}).centroid.y

joined = gpd.sjoin(plot_utm, tile_idx[['geometry', 'file_name', 'acq_name']]).drop('index_right', axis=1)

joined.info()

joined.reset_index().rename({'index':'uuid'}, axis=1).head()

# we'll need to do a groupby forest data source (USFS, BLM, DNR) 
# as well as lidar acquisition name so that each combo can be processed separately
# we'll need to do this for both UTM 10 and 11

GROUP_COLS = ['acq_name', 'source', 'uuid', 'file_name']
matching_tiles = joined.reset_index().rename({'index':'uuid'}, axis=1).groupby(by=GROUP_COLS).sum().reset_index()[GROUP_COLS].set_index('uuid')

plot_utm = plot_utm.loc[joined.index]

@dask.delayed
def clip_1ha_plots(uuid):
    source = plot_utm.loc[uuid]['source']
    TILE_DIR = '../data/raw/lidar/noaa_tiles/utm_laz/'
    
    if 'BLM' in source:
        ODIR = '../data/interim/lidar/plot_clips/noaa/blm_plots/hectare_clips/'
    elif 'USFS' in source:
        ODIR = '../data/interim/lidar/plot_clips/noaa/usfs_plots/hectare_clips/'
    elif 'WA-DNR' in source:
        ODIR = '../data/interim/lidar/plot_clips/noaa/dnr_plots/hectare_clips/'
    
    # grab the coordinates of the plot for clipping
    x, y = plot_utm.loc[uuid][['utm_x', 'utm_y']]
    
    PLOT_RADIUS_M = 56.418  # radius of 1-hectare plot (185.1 ft)
    
    # find all matching tiles 
    all_matching_tiles = matching_tiles.loc[[uuid]]
    # loop through each lidar acquisition to clip the plots
    procs = []
    for acq in pd.unique(all_matching_tiles['acq_name']):
        out_name = uuid + '_' + acq + '.laz'
        outfile = os.path.join(ODIR, out_name)
        
        tiles = all_matching_tiles.loc[all_matching_tiles['acq_name'] == acq]['file_name'].values
        tiles = [os.path.join(TILE_DIR, t) for t in tiles]
        print(tiles)

        proc = subprocess.run(['wine', '/storage/lidar/LAStools/bin/las2las.exe',
                               '-i', *tiles, '-merged',
                               '-o', outfile,
                               '-keep_circle', str(x), str(y), str(PLOT_RADIUS_M)],
                              stderr=subprocess.PIPE,
                              stdout=subprocess.PIPE)
        procs.append(proc)
#         print(outfile, flush=True)
#     print(uuid, flush=True)
    return procs

@dask.delayed
def clip_smaller_plots(ha_clip):
    dirname, basename = os.path.split(ha_clip)
    outfile = os.path.abspath(os.path.join(dirname, '..', 'plot_clips', basename))

    uuid = os.path.basename(ha_clip).split('_')[0]
    
    # grab the coordinates of the plot for clipping
    x, y = plot_utm.loc[uuid][['utm_x', 'utm_y']]
    
    source = plot_utm.loc[uuid]['source']
    if 'WA-DNR' in source:
        # radius of 1/10th acre plot (37.2 ft)
        PLOT_RADIUS_M = 11.35  
    elif 'USFS' in source:
        # radius of 1/4 acre plot (58.9 ft)
        PLOT_RADIUS_M = 17.9454 
    elif 'BLM' in source:
        # radius of 1/8th acre plot (41.6 ft)
        PLOT_RADIUS_M = 12.69  
    
    proc = subprocess.run(['wine', '/storage/lidar/LAStools/bin/las2las.exe',
                           '-i', ha_clip,
                           '-o', outfile,
                           '-keep_circle', str(x), str(y), str(PLOT_RADIUS_M)],
                          stderr=subprocess.PIPE,
                          stdout=subprocess.PIPE)
#     print(outfile, flush=True)
    return proc

cluster=LocalCluster(scheduler_port=7001, diagnostics_port=7002)
c = Client(cluster)

c.scatter([plot_utm, joined, matching_tiles])

plot_ids = pd.unique(joined.index)
len(plot_ids)

res = c.persist([clip_1ha_plots(i) for i in plot_ids])
progress(res)

usfs_ha = glob.glob('../data/interim/lidar/plot_clips/noaa/usfs_plots/hectare_clips/*.laz') 
blm_ha = glob.glob('../data/interim/lidar/plot_clips/noaa/blm_plots/hectare_clips/*.laz')
dnr_ha = glob.glob('../data/interim/lidar/plot_clips/noaa/dnr_plots/hectare_clips/*.laz')
ha_clips = usfs_ha + blm_ha + dnr_ha

res = c.persist([clip_smaller_plots(f) for f in ha_clips])
progress(res)

usfs_plot = glob.glob('../data/interim/lidar/plot_clips/noaa/usfs_plots/plot_clips/*.laz') 
blm_plot = glob.glob('../data/interim/lidar/plot_clips/noaa/blm_plots/plot_clips/*.laz')
dnr_plot = glob.glob('../data/interim/lidar/plot_clips/noaa/dnr_plots/plot_clips/*.laz')
plot_clips = usfs_plot + blm_plot + dnr_plot
len(plot_clips)

@dask.delayed
def get_boundary(infile, odir):
    proc = subprocess.run(['wine', '/storage/lidar/LAStools/bin/lasboundary.exe',
                           '-i', infile,
                           '-odir', odir,
                           '-oshp',
                           '-labels'], 
                          stderr=subprocess.PIPE,
                          stdout=subprocess.PIPE)
#     print(outfile)
    return proc

USFS_HA = glob.glob('../data/interim/lidar/plot_clips/noaa/usfs_plots/hectare_clips/*.laz')
USFS_PLOT = glob.glob('../data/interim/lidar/plot_clips/noaa/usfs_plots/plot_clips/*.laz')
DNR_HA = glob.glob('../data/interim/lidar/plot_clips/noaa/dnr_plots/hectare_clips/*.laz')
DNR_PLOT = glob.glob('../data/interim/lidar/plot_clips/noaa/dnr_plots/plot_clips/*.laz')
BLM_HA = glob.glob('../data/interim/lidar/plot_clips/noaa/blm_plots/hectare_clips/*.laz')
BLM_PLOT = glob.glob('../data/interim/lidar/plot_clips/noaa/blm_plots/plot_clips/*.laz')

jobs_to_do = []
PLOT_SETS = [USFS_HA, USFS_PLOT, DNR_HA, DNR_PLOT, BLM_HA, BLM_PLOT]
ODIRS = ['../data/interim/lidar/plot_clips/noaa/usfs_plots/hectare_clips/',
         '../data/interim/lidar/plot_clips/noaa/usfs_plots/plot_clips/',
         '../data/interim/lidar/plot_clips/noaa/dnr_plots/hectare_clips/',
         '../data/interim/lidar/plot_clips/noaa/dnr_plots/plot_clips/',
         '../data/interim/lidar/plot_clips/noaa/blm_plots/hectare_clips/',
         '../data/interim/lidar/plot_clips/noaa/blm_plots/plot_clips/']

for plot_set, odir in zip(PLOT_SETS, ODIRS):
    jobs_to_do += [get_boundary(p, odir=odir) for p in plot_set]

res = c.persist(jobs_to_do)
progress(res)

# c.cancel(res)

# c.close()
# cluster.close()
