==> notebooks/11A_Combining Features for Plots.ipynb <==
import numpy as np
import pandas as pd
import geopandas as gpd
from matplotlib import pyplot as plt
import seaborn as sns
sns.set_style('darkgrid')
%matplotlib inline

pd.set_option('display.max_columns', None)  
pd.set_option('display.max_info_columns', 999)

# INVENTORY FEATURES
FVS = '../data/processed/fvs_outputs/fvs_outputs.csv'

# CLIMATE FEATURES
CLIMATE = '../data/processed/pnw_lidar_plots_climatewna_2000-2017MSY.csv'

# LIDAR FEATURES
CLOUDMETS = '../data/processed/lidar/plot_clips/cloudmetrics.csv'
GRIDSURF = '../data/processed/lidar/plot_clips/gridsurfacestats.csv'
TOPOMETS = '../data/processed/lidar/plot_clips/topometrics.csv'

# SATELLITE FEATURES
L8 = '../data/interim/gee/landsat_2014-2020.csv'
# MODIS = '../data/interim/modis_metrics.csv'
S2 = '../data/interim/gee/sentinel2sr_2019-2020.csv'
S1 = '../data/interim/gee/sentinel1_2015-2020.csv'
PALSAR = '../data/interim/gee/palsar_2010-2018.csv'
LANDTRENDR = '../data/interim/gee/landtrendr_2010-2020.csv'

## Combine features that are time-invariant for a plot
We'll identify the soil types, distance from water, and some attributes like latitude, longitude, potential vegetation type, and ecoregion.

# PLOT FEATURES
PLOT_LOCS = '../data/processed/blm_usfs_wadnr_plot_footprints.shp'
SOILS = '../data/processed/soils/soils_data_forPlots.csv'
POTVEG = '../data/processed/pot_veg/pot_veg_plots.csv'
WATER = '../data/interim/dist_to_water/water_distance_fromPlots.csv'
ECOREGIONS = '../data/raw/epa/us_eco_l3_state_boundaries.shp'

plot_locs = gpd.read_file(PLOT_LOCS).to_crs(5070)
KEEP_PLOT_COLS = ['lat', 'lon', 'orig_id', 'uuid', 'source', 'meas_yr', 'meas_date', 'geometry']
plot_locs['uuid'] = plot_locs.uuid.str[0:8]
plot_locs = plot_locs[KEEP_PLOT_COLS]
plot_locs.info()

We need to do a spatial join to identify the ecoregion that each plot falls within

ecoreg = gpd.read_file(ECOREGIONS)
ecoreg = ecoreg.loc[ecoreg.STATE_NAME.isin(['Washington', 'Oregon'])]
ecoreg['ecoregion3'] = ecoreg.US_L3NAME.str.lower().str.replace(' ', '_').str.replace('/', '_')
ecoreg.crs = {'init': 'epsg:5070'}
ecoreg.info()

# add the ecoregion to the plot table
plot_locs['geometry'] = plot_locs.geometry.centroid
ecoreg_cols = ecoreg.columns.drop('geometry')
ecoreg_join = gpd.sjoin(plot_locs.to_crs(5070), ecoreg.to_crs(5070), how='left').drop(['index_right'], axis=1)
ecoreg_join.loc[ecoreg_join.uuid == '9d4ab655', ecoreg_cols] = ecoreg.loc[ecoreg.ecoregion3 == 'puget_lowland', ecoreg_cols].drop_duplicates().values

KEEP_PLOT_COLS = ['lat', 'lon', 'orig_id', 'uuid', 'source', 'meas_yr', 'ecoregion3']
plot_info = ecoreg_join[KEEP_PLOT_COLS].copy()

plot_info.loc[plot_info.source.str.contains('USFS'), 'agency'] = 'USFS'
plot_info.loc[plot_info.source.str.contains('BLM'), 'agency'] = 'BLM'
plot_info.loc[plot_info.source.str.contains('WA-DNR'), 'agency'] = 'WADNR'

plot_info.loc[plot_info.agency == 'USFS', 'plot_size_ac'] = 1/4
plot_info.loc[plot_info.agency == 'BLM', 'plot_size_ac'] = 1/8
plot_info.loc[plot_info.agency == 'WADNR', 'plot_size_ac'] = 1/10

plot_info = plot_info.set_index('uuid')
plot_info.info()

soils = pd.read_csv(SOILS)
soils['uuid'] = soils.uuid.str[0:8]
soils = soils.set_index('uuid')
soils.columns = [col.lower() for col in soils.columns]
KEEP_SOILS = ['bulk_dens', 'soil_depth', 'pct_clay_surf', 'pct_rock_surf', 'pct_sand_surf']
soils = soils[KEEP_SOILS]
soils.info()

potveg = pd.read_csv(POTVEG)
potveg['uuid'] = potveg['uuid'].str[0:8]
potveg = potveg.set_index('uuid')
potveg.columns = [col.lower() for col in potveg.columns]
potveg = potveg.rename({'esp2':'pot_veg_type'}, axis=1)[['pot_veg_type']]
potveg.info()

water = pd.read_csv(WATER)
water['uuid'] = water.uuid.str[0:8]
water = water.set_index('uuid')
water.info()

plot_atts = plot_info.join(soils, how='outer').join(water, how='outer').join(potveg, how='outer')
plot_atts.info()

plot_atts.to_csv('../data/processed/plot_features.csv', index=True, header=True)

## Prepare monthly, seasonal, and annual climate features

climate = pd.read_csv(CLIMATE).rename(
    {'Year': 'year', 'ID1':'uuid', 'ID2':'source'},
    axis=1)

climate = climate.set_index(['uuid', 'year'])

# breaking the climatena columns into groups
MONTHLY_COLS = ['Tmax01','Tmax02','Tmax03','Tmax04','Tmax05','Tmax06',
                'Tmax07','Tmax08','Tmax09','Tmax10','Tmax11','Tmax12',
                'Tmin01','Tmin02','Tmin03','Tmin04','Tmin05','Tmin06',
                'Tmin07','Tmin08','Tmin09','Tmin10','Tmin11','Tmin12',
                'Tave01','Tave02','Tave03','Tave04','Tave05','Tave06',
                'Tave07','Tave08','Tave09','Tave10','Tave11','Tave12',
                'PPT01','PPT02','PPT03','PPT04','PPT05','PPT06',
                'PPT07','PPT08','PPT09','PPT10','PPT11','PPT12',
                'Rad01','Rad02','Rad03','Rad04','Rad05','Rad06',
                'Rad07','Rad08','Rad09','Rad10','Rad11','Rad12',
                'DD_0_01','DD_0_02','DD_0_03','DD_0_04','DD_0_05','DD_0_06',
                'DD_0_07','DD_0_08','DD_0_09','DD_0_10','DD_0_11','DD_0_12',
                'DD5_01','DD5_02','DD5_03','DD5_04','DD5_05','DD5_06',
                'DD5_07','DD5_08','DD5_09','DD5_10','DD5_11','DD5_12',
                'DD_18_01','DD_18_02','DD_18_03','DD_18_04','DD_18_05','DD_18_06',
                'DD_18_07','DD_18_08','DD_18_09','DD_18_10','DD_18_11','DD_18_12',
                'DD18_01','DD18_02','DD18_03','DD18_04','DD18_05','DD18_06',
                'DD18_07','DD18_08','DD18_09','DD18_10','DD18_11','DD18_12',
                'NFFD01','NFFD02','NFFD03','NFFD04','NFFD05','NFFD06',
                'NFFD07','NFFD08','NFFD09','NFFD10','NFFD11','NFFD12',
                'PAS01','PAS02','PAS03','PAS04','PAS05','PAS06',
                'PAS07','PAS08','PAS09','PAS10','PAS11','PAS12',
                'Eref01','Eref02','Eref03','Eref04','Eref05','Eref06',
                'Eref07','Eref08','Eref09','Eref10','Eref11','Eref12',
                'CMD01','CMD02','CMD03','CMD04','CMD05','CMD06',
                'CMD07','CMD08','CMD09','CMD10','CMD11','CMD12',
                'RH01','RH02','RH03','RH04','RH05','RH06',
                'RH07','RH08','RH09','RH10','RH11','RH12']
SEASON_COLS = ['Tmax_wt','Tmax_sp','Tmax_sm','Tmax_at',
               'Tmin_wt','Tmin_sp','Tmin_sm','Tmin_at',
               'Tave_wt','Tave_sp','Tave_sm','Tave_at',
               'PPT_wt','PPT_sp','PPT_sm','PPT_at',
               'Rad_wt','Rad_sp','Rad_sm','Rad_at',
               'DD_0_wt','DD_0_sp','DD_0_sm','DD_0_at',
               'DD5_wt','DD5_sp','DD5_sm','DD5_at',
               'DD_18_wt','DD_18_sp','DD_18_sm','DD_18_at',
               'DD18_wt','DD18_sp','DD18_sm','DD18_at',
               'NFFD_wt','NFFD_sp','NFFD_sm','NFFD_at',
               'PAS_wt','PAS_sp','PAS_sm','PAS_at',
               'Eref_wt','Eref_sp','Eref_sm','Eref_at',
               'CMD_wt','CMD_sp','CMD_sm','CMD_at',
               'RH_wt','RH_sp','RH_sm','RH_at']
ANN_COLS = ['MAT','MWMT','MCMT','TD','MAP','MSP',
            'AHM','SHM','DD_0','DD5','DD_18','DD18',
            'NFFD','bFFP','eFFP','FFP','PAS','EMT',
            'EXT','MAR','Eref','CMD','RH']

climate_monthly = climate[MONTHLY_COLS].replace(-9999.0, np.nan)
climate_seasonal = climate[SEASON_COLS].replace(-9999.0, np.nan).drop(['Rad_wt','Rad_sp','Rad_sm','Rad_at'], axis=1)
climate_annual = climate[ANN_COLS].replace(-9999.0, np.nan).drop('MAR', axis=1)

# replace current year values for fall and winter with previous year values
# to reflect run-up to current growing season
FALL_WINT_COLS = [x for x in climate_monthly.columns if x[-2:] in ['09', '10', '11', '12']]
climate_monthly[FALL_WINT_COLS] = climate_monthly.groupby(level=[0,1])[FALL_WINT_COLS].sum().shift()
climate_monthly.loc[climate_monthly.index.get_level_values(1) == 2000, FALL_WINT_COLS] = np.nan

climate_annual.head()

climate_monthly.to_csv('../data/processed/climate_features_monthly.csv', index=True, header=True)
climate_seasonal.to_csv('../data/processed/climate_features_seasonal.csv', index=True, header=True)
climate_annual.to_csv('../data/processed/climate_features_annual.csv', index=True, header=True)

## Combine LiDAR Features
We've got cloudmetrics, gridsurfacestats, and topometrics attributes to combine

cloudmets = pd.read_csv(CLOUDMETS, low_memory=False)
cloudmets['uuid'] = cloudmets.FileTitle.apply(lambda x: x.split('_')[0]).str[0:8]
cloudmets['lidar_year'] = cloudmets.FileTitle.apply(lambda x: int(x.split('_')[-1].split('-')[-1]))
cloudmets['lidar_acq'] = cloudmets.FileTitle.apply(lambda x: '_'.join(x.split('_')[1:3]))

RENAME_COLS = {'Total return count': 'num_returns',
               'Elev strata (below 0.15) return proportion':'strat0_return-proportion',
               'Elev strata (0.15 to 1.37) return proportion':'strat1_return-proportion',
               'Elev strata (5.00 to 10.00) return proportion':'strat2_return-proportion',
               'Elev strata (10.00 to 20.00) return proportion':'strat3_return-proportion',
               'Elev strata (20.00 to 30.00) return proportion':'strat4_return-proportion',
               'Elev strata (above 30.00) return proportion':'strat5_return-proportion',
               'Elev skewness': 'skewness',
               'Elev kurtosis': 'kurtosis',
               'Int strata (below 0.15) median':'strat0_intensity-median',
               'Int strata (0.15 to 1.37) median':'strat1_intensity-median',
               'Int strata (1.37 to 5.00) median':'strat2_intensity-median',
               'Int strata (5.00 to 10.00) median':'strat3_intensity-median',
               'Int strata (10.00 to 20.00) median':'strat4_intensity-median',
               'Int strata (above 30.00) median':'strat5_intensity-median',
               'Elev P05':'height_05-percentile','Elev P25':'height_25-percentile',
               'Elev P50':'height_50-percentile',
               'Elev P75':'height_75-percentile',
               'Elev P95':'height_95_percentile',
               'Elev maximum':'height_max',
               'Percentage all returns above 1.37':'cover'}

cloudmets = cloudmets.rename(RENAME_COLS, axis=1)
KEEP_COLS = [col for col in RENAME_COLS.values()] + ['uuid', 'lidar_acq', 'lidar_year']
cloudmets = cloudmets[KEEP_COLS].set_index(['uuid', 'lidar_year'])
cloudmets.info()

cloudmets.head()

gridsurf = pd.read_csv(GRIDSURF)
gridsurf['uuid'] = gridsurf.plot_id.apply(lambda x: x.split('_')[0]).str[0:8]
gridsurf['lidar_year'] = gridsurf.plot_id.apply(lambda x: int(x.split('_')[-1].split('-')[-1]))
KEEP_COLS = ['uuid', 'lidar_year', 'potential_volume', 'stddev_height',
             'surface_area_ratio', 'surface_volume', 'surface_volume_ratio']
gridsurf = gridsurf[KEEP_COLS].set_index(['uuid', 'lidar_year'])
gridsurf.info()

gridsurf.head()

topomets = pd.read_csv(TOPOMETS)
topomets['uuid'] = topomets.plot_id.apply(lambda x: x.split('_')[0]).str[0:8]
topomets['lidar_year'] = topomets.plot_id.apply(lambda x: int(x.split('_')[-1].split('-')[-1]))
topomets = topomets.drop('plot_id', axis=1).set_index(['uuid', 'lidar_year'])
topomets.info()

lidar_features = cloudmets.join(gridsurf, how='outer').join(topomets, how='outer')
lidar_features.info()

lidar_features.to_csv('../data/processed/lidar_features.csv', index=True, header=True)

## Inventory features

fvs = pd.read_csv(FVS)
fvs['uuid'] = fvs['uuid'].str[0:8]
fvs = fvs.set_index(['uuid', 'year'])
fvs.info()

fvs.to_csv('../data/processed/inventory_features.csv', index=True, header=True)

## Combine satellite-derived features
We've got data from Sentinel-1, Sentinel-2, LANDSAT 8, PALSAR, and Landtrendr time-series.

s1 = pd.read_csv(S1)
s1['uuid'] = s1['uuid'].str[0:8]
s1 = s1.set_index(['uuid', 'year'])
s1.columns = ['_'.join(['S1', col.upper()]) for col in s1.columns]
s1.info()

s2 = pd.read_csv(S2)
s2['uuid'] = s2['uuid'].str[0:8]
s2 = s2.set_index(['uuid', 'year'])
s2.columns = ['_'.join(['S2', col.upper()]) for col in s2.columns]
s2.info()

landsat = pd.read_csv(L8)
landsat['uuid'] = landsat['uuid'].str[0:8]
landsat = landsat.set_index(['uuid', 'year'])
landsat.columns = ['_'.join(['L8', col.upper()]) for col in landsat.columns]
landsat = landsat/1000.
landsat.info()

palsar = pd.read_csv(PALSAR)
palsar['uuid'] = palsar['uuid'].str[0:8]
palsar = palsar.set_index(['uuid', 'year'])
palsar = palsar[['HH', 'HV']]
palsar.columns = ['_'.join(['P', col.upper()]) for col in palsar.columns]
palsar.info()

landtrendr = pd.read_csv(LANDTRENDR)
landtrendr['uuid'] = landtrendr['uuid'].str[0:8]
landtrendr = landtrendr.set_index(['uuid', 'year'])
landtrendr.columns = ['_'.join(['LT', col.upper()]) for col in landtrendr.columns]
landtrendr = landtrendr.astype(int)
landtrendr.info()

landtrendr.columns

satellite_features = s1.join(s2, how='outer').join(palsar, how='outer').join(landsat, how='outer').join(landtrendr, how='outer')

df = satellite_features
for season in ['LEAFON', 'LEAFOFF']:
    for sensor in ['L8', 'S2']:
        R, G, B = f'{sensor}_R_{season}', f'{sensor}_G_{season}', f'{sensor}_B_{season}'
        NIR, SWIR1, SWIR2 =  f'{sensor}_NIR_{season}', f'{sensor}_SWIR1_{season}', f'{sensor}_SWIR2_{season}'

        NDVI = f'{sensor}_NDVI_{season}'
        df[NDVI] = (df[NIR] - df[R])/(df[NIR] + df[R])

        ENDVI = f'{sensor}_ENDVI_{season}'
        df[NDVI] = (df[NIR] + df[G] - 2*df[B])/(df[NIR] + df[G] + 2*df[B])

        SAVI = f'{sensor}_SAVI_{season}'
        df[SAVI] = 1.5*(df[NIR] - df[R])/(df[NIR] + df[R] + 0.5)
        
        BRIGHTNESS = f'{sensor}_BRIGHTNESS_{season}'
        df[BRIGHTNESS] = 0.3029*df[B] + 0.2786*df[G] + 0.4733*df[R] + 0.5599*df[NIR] + 0.508*df[SWIR1] + 0.1872*df[SWIR2]
        
        GREENNESS = f'{sensor}_GREENNESS_{season}'
        df[GREENNESS] = -0.2941*df[B] + -0.243*df[G] + -0.5424*df[R] + 0.7276*df[NIR] + 0.0713*df[SWIR1] + -0.1608*df[SWIR2]
        
        WETNESS = f'{sensor}_WETNESS_{season}'
        df[WETNESS] = 0.1511*df[B] + 0.1973*df[G] + 0.3283*df[R] + 0.3407*df[NIR] + -0.7117*df[SWIR1] + -0.4559*df[SWIR2]

for sensor in ['L8', 'S2']:
    if sensor == 'S2':
        addl_bands = ['RE1', 'RE2', 'RE3', 'RE4']
    else:
        addl_bands =  []
    for band in ['R', 'G', 'B', 'NIR', 'SWIR1', 'SWIR2', 'NDVI', 'SAVI', 'BRIGHTNESS', 'GREENNESS', 'WETNESS'] + addl_bands:
        df[f'{sensor}_d{band}'] = df[f'{sensor}_{band}_LEAFON'] - df[f'{sensor}_{band}_LEAFOFF']

satellite_features.info()

satellite_features.to_csv('../data/processed/satellite_features.csv', index=True, header=True)
