==> notebooks/19_Visualizing Training Data.ipynb <==
import numpy as np
import pandas as pd
import geopandas as gpd
from matplotlib import pyplot as plt
import seaborn as sns
sns.set_style('darkgrid')
%matplotlib inline

PLOT_LOCS = '../data/processed/blm_usfs_wadnr_plot_footprints.shp'
ECOREGIONS = '../data/raw/epa/us_eco_l3_state_boundaries.shp'

plot_locs = gpd.read_file(PLOT_LOCS)
plot_locs['uuid_part'] = plot_locs['uuid'].apply(lambda x: x.split('-')[0])
plot_locs['geometry'] = plot_locs.centroid

ecoreg = gpd.read_file(ECOREGIONS)
ecoreg = ecoreg.loc[ecoreg.STATE_NAME.isin(['Washington', 'Oregon'])]
ecoreg['ecoregion3'] = ecoreg.US_L3NAME.str.lower().str.replace(' ', '_').str.replace('/', '_')
ecoreg.crs = {'init': 'epsg:5070'}
ecoreg.info()

LIDAR_DATA = '../data/processed/lidar_structure_training_data.csv'
SENTINEL_DATA = '../data/processed/sentinel_structure_training_data.csv'

lid_train = pd.read_csv(LIDAR_DATA)
sat_train = pd.read_csv(SENTINEL_DATA)

lid_train.info()

lid_train.columns

lid_train.head()

lid_pivot = pd.pivot_table(lid_train.groupby(by=['ecoregion3', 'year'])[['uuid', 'lidar_acq']].nunique().reset_index(), index=['ecoregion3'], columns=['year'], values=['uuid'])['uuid']
lid_pivot.index = [' '.join(idx.split('_')[0:2]).upper() for idx in lid_pivot.index]
lid_pivot.columns = [int(col) for col in lid_pivot.columns]
lid_pivot = lid_pivot.drop('COLUMBIA PLATEAU')
lid_pivot['TOTAL'] = lid_pivot.sum(axis=1).astype(int)
lid_pivot.loc['TOTAL'] = lid_pivot.sum(axis=0).astype(int)
lid_pivot.fillna('--')

lid_train['uuid'].nunique()

sat_train.groupby(by=['ecoregion3'])['uuid'].nunique()

lid_pivot.shape

sat_pivot = pd.pivot_table(sat_train.groupby(by=['ecoregion3', 'year'])[['uuid']].nunique().reset_index(), index=['ecoregion3'], columns=['year'], values=['uuid'])['uuid']
sat_pivot.index = [' '.join(idx.split('_')[0:2]).upper() for idx in sat_pivot.index]
sat_pivot.columns = [int(col) for col in sat_pivot.columns]
sat_pivot = sat_pivot.drop('COLUMBIA PLATEAU')
sat_pivot['TOTAL'] = sat_pivot.sum(axis=1).astype(int)
sat_pivot.loc['TOTAL'] = sat_pivot.sum(axis=0).astype(int)
sat_pivot.fillna('--')

len(pd.unique(lid_train.uuid)), len(pd.unique(sat_train.uuid))

USE_LIDAR_COLS = ['strat0_return-proportion', 'strat1_return-proportion', 
                  'strat2_return-proportion', 'strat3_return-proportion', 
                  'strat4_return-proportion', 'strat5_return-proportion', 
                  'height_05-percentile',  'height_25-percentile', 
                  'height_50-percentile', 'height_75-percentile',
                  'height_95_percentile', 'cover', 
                  'potential_volume', 'stddev_height', 
                  'surface_volume', 'kurtosis', 'skewness']
LIDAR_X_COLS = USE_LIDAR_COLS + ['elevation', 'lat', 'lon']
len(LIDAR_X_COLS) - 3

SAT_COLS = ['S2_R_LEAFOFF', 'S2_G_LEAFOFF', 'S2_B_LEAFOFF', 'S2_NIR_LEAFOFF', 'S2_SWIR1_LEAFOFF',
            'S2_SWIR2_LEAFOFF', 'S2_RE1_LEAFOFF', 'S2_RE2_LEAFOFF', 'S2_RE3_LEAFOFF', 
            'S2_R_LEAFON', 'S2_G_LEAFON', 'S2_B_LEAFON', 'S2_NIR_LEAFON', 'S2_SWIR1_LEAFON',
            'S2_SWIR2_LEAFON', 'S2_RE1_LEAFON', 'S2_RE2_LEAFON', 'S2_RE3_LEAFON', 'S2_RE4_LEAFON',
            'S2_NDVI_LEAFON', 'S2_SAVI_LEAFON', 'S2_BRIGHTNESS_LEAFON', 'S2_GREENNESS_LEAFON',
            'S2_WETNESS_LEAFON', 'S2_NDVI_LEAFOFF', 'S2_SAVI_LEAFOFF', 'S2_BRIGHTNESS_LEAFOFF',
            'S2_GREENNESS_LEAFOFF', 'S2_WETNESS_LEAFOFF', 'S2_dR', 'S2_dG', 'S2_dB', 'S2_dNIR',
            'S2_dSWIR1', 'S2_dSWIR2', 'S2_dRE1', 'S2_dRE2', 'S2_dNDVI', 'S2_dSAVI', 'S2_dBRIGHTNESS',
            'S2_dGREENNESS', 'S2_dWETNESS', 'S2_dRE3', 'S2_dRE4', 
            'LT_DUR_NBR', 'LT_DUR_SWIR1', 'LT_MAG_NBR', 'LT_MAG_SWIR1', 'LT_RATE_NBR', 'LT_RATE_SWIR1',
            'LT_YSD_NBR', 'LT_YSD_SWIR1', 
            'elevation', 'lat', 'lon']

len(SAT_COLS) - 3
